Account for case weights in survival and cumulative hazard KM plots
The Kaplan Meier produced by `plot.flexsurvreg` do not account for case weights (when used). While this issue can be easily worked around, the plots can be misleading as all cases are associated to unit weights, and weighted KM are expected.

As all `flexsurvreg` objects contain case weights in `$data$m$`(weights)`` this can be added easily to the `type == "survival"` and `type == "cumhaz"` options in `plot.flexsurvreg`. However, the `muhaz `function used in `type == "hazard"` does not have a built-in functionality accounting for case weights.